% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dStoN.R
\name{dStoN}
\alias{dStoN}
\title{Probability Density Function of Stomped Normal (StoN) Distribution}
\usage{
dStoN(x, mu, sig, k)
}
\arguments{
\item{x}{Input random variable or vector.}

\item{mu}{Parameter \eqn{\mu} of the distribution.}

\item{sig}{Parameter \eqn{\sigma} of the distribution.}

\item{k}{Parameter \emph{k} of the distribution.}
}
\value{
The PDF of StoN with parameters(mu,sigma,k) at X=x.
}
\description{
Probability Density Function of Stomped Normal (StoN) Distribution
}
\examples{
dStoN(1,0,5,1)
dStoN(c(1,5,4),0,1,0.5)
}
