% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.mle.R
\name{estimate.mle}
\alias{estimate.mle}
\title{Likelihood Iterative Estimators (LIEs) of the Parameters of Stomped Normal (StN) Distribution}
\usage{
estimate.mle(XS)
}
\arguments{
\item{XS}{Input dataset.}
}
\value{
LI Estimators of Parameters (mu, sig, k).
}
\description{
Likelihood Iterative Estimators (LIEs) of the Parameters of Stomped Normal (StN) Distribution
}
\examples{
estimate.mle(rStoN(1000,0,5,1))
}
