% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.mom.R
\name{estimate.mom}
\alias{estimate.mom}
\title{Method of Moments Estimators of the Parameters of Stomped Normal (StN) Distribution}
\usage{
estimate.mom(XS)
}
\arguments{
\item{XS}{Input dataset.}
}
\value{
MoM Estimators of Parameters (mu, sig, k).
}
\description{
Method of Moments Estimators of the Parameters of Stomped Normal (StN) Distribution
}
\examples{
estimate.mom(rStoN(1000,0,5,1))
}
