% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLestimate.k.R
\name{MLestimate.k}
\alias{MLestimate.k}
\title{MLE of Width Parameter \emph{k} of Stomped Normal (StN) Distribution}
\usage{
MLestimate.k(data, mu, sig)
}
\arguments{
\item{data}{Input random vector.}

\item{mu}{Parameter \eqn{\mu} of the distribution.}

\item{sig}{Parameter \eqn{\sigma} of the distribution.}
}
\value{
MLE of Parameter \emph{k} of the distribution.
}
\description{
MLE of Width Parameter \emph{k} of Stomped Normal (StN) Distribution
}
\examples{
MLestimate.k(rStoN(1000,0,5,1),0,5)
}
