% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pStoN.R
\name{pStoN}
\alias{pStoN}
\title{Cumulative Density Function of Stomped Normal (StoN) Distribution}
\usage{
pStoN(x, mu, sig, k)
}
\arguments{
\item{x}{Input random variable or vector.}

\item{mu}{Parameter \eqn{\mu} of the distribution.}

\item{sig}{Parameter \eqn{\sigma} of the distribution.}

\item{k}{Parameter \emph{k} of the distribution.}
}
\value{
The CDF of StoN with parameters(mu,sigma,k) at X=x.
}
\description{
Cumulative Density Function of Stomped Normal (StoN) Distribution
}
\examples{
pStoN(1,0,5,1)
pStoN(c(1,5,4),0,1,0.5)
}
