% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rStoN.R
\name{rStoN}
\alias{rStoN}
\title{Random Sample Generation from Stomped Normal (StoN) Distribution}
\usage{
rStoN(n, mu, sig, k)
}
\arguments{
\item{n}{The number of observations.}

\item{mu}{Parameter \eqn{\mu} of the distribution.}

\item{sig}{Parameter \eqn{\sigma} of the distribution.}

\item{k}{Parameter \emph{k} of the distribution.}
}
\value{
The random sample of size n from StoN with parameters(mu,sigma,k).
}
\description{
Random Sample Generation from Stomped Normal (StoN) Distribution
}
\examples{
rStoN(1000,0,5,1)
}
